+++ src/main/java/cn/simplifydb/database/run/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 数据库执行相关类
+ */
+package cn.simplifydb.database.run;